# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "SmartScheduler API",
    "description": "SmartScheduler API for job scheduling and contractor management",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5004"
    }
  ],
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get all jobs",
        "description": "Retrieves a list of all jobs. Optionally filter by status and priority, and limit results.",
        "operationId": "GetJobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a new job",
        "description": "Creates a new job with the provided information. Address validation and timezone lookup are performed automatically.",
        "operationId": "CreateJob",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job by ID",
        "description": "Retrieves a specific job by its unique identifier.",
        "operationId": "GetJobById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update an existing job",
        "description": "Updates an existing job with the provided information. Only provided fields will be updated.",
        "operationId": "UpdateJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}/latest-recommendations": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get latest recommendations for a job",
        "description": "Retrieves the most recent contractor recommendations for a job. Returns 404 if no recommendations have been calculated.",
        "operationId": "GetLatestRecommendations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}/status": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update job status",
        "description": "Updates the status of an existing job. Valid status transitions are enforced.",
        "operationId": "UpdateJobStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobStatusRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}/assign": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Assign job to contractor",
        "description": "Assigns a job to a contractor with the specified time slot. Availability is re-validated before assignment. Returns existing assignment if same request (idempotent).",
        "operationId": "AssignJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignJobRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}/reschedule": {
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Reschedule job",
        "description": "Reschedules a job to a new time slot. Validates feasibility for all assigned contractors and updates assignments accordingly.",
        "operationId": "RescheduleJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleJobRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        }
      }
    },
    "/api/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel job",
        "description": "Cancels a job and all its active assignments. Cannot cancel completed jobs.",
        "operationId": "CancelJob",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelJobRequest"
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoLocationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "formattedAddress": {
            "type": "string",
            "nullable": true
          },
          "placeId": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "JobDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocationDto"
          },
          "timezone": {
            "type": "string"
          },
          "requiredSkills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "serviceWindow": {
            "$ref": "#/components/schemas/TimeWindowDto"
          },
          "priority": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "assignmentStatus": {
            "type": "string"
          },
          "assignedContractors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContractorAssignmentDto"
            }
          },
          "accessNotes": {
            "type": "string",
            "nullable": true
          },
          "tools": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "desiredDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastRecommendationAuditId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "TimeWindowDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ContractorAssignmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "distanceMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "travelTimeMinutes": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "RecommendationResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string",
            "format": "guid"
          },
          "jobId": {
            "type": "string",
            "format": "guid"
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecommendationDto"
            }
          },
          "bestRecommendationContractorId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "configVersion": {
            "type": "integer",
            "format": "int32"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RecommendationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "contractorName": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "scoreBreakdown": {
            "$ref": "#/components/schemas/ScoreBreakdown"
          },
          "rationale": {
            "type": "string"
          },
          "suggestedSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeSlotDto"
            }
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "eta": {
            "type": "integer",
            "format": "int32"
          },
          "contractorBaseLocation": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CreateJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocationDto"
          },
          "placeId": {
            "type": "string",
            "nullable": true
          },
          "serviceWindow": {
            "$ref": "#/components/schemas/TimeWindowDto"
          },
          "priority": {
            "type": "string"
          },
          "requiredSkills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessNotes": {
            "type": "string",
            "nullable": true
          },
          "tools": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "desiredDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "location": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/GeoLocationDto"
              }
            ]
          },
          "placeId": {
            "type": "string",
            "nullable": true
          },
          "serviceWindow": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TimeWindowDto"
              }
            ]
          },
          "priority": {
            "type": "string",
            "nullable": true
          },
          "requiredSkills": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "accessNotes": {
            "type": "string",
            "nullable": true
          },
          "tools": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "desiredDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "UpdateJobStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        }
      },
      "JobStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Scheduled",
          "InProgress",
          "Completed",
          "Canceled"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
      "AssignmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobId": {
            "type": "string",
            "format": "guid"
          },
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "auditId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssignJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          },
          "auditId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "source": {
            "type": "string"
          }
        }
      },
      "RescheduleJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "startUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CancelJobRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}